Fix intersections for literals and imports (#87)
* Implement support for literal type object intersections

* Add test for exported intersection type w/ type aliases

* Fix intersection with empty object literals

* Fix test - should have failed

* Fix test export-intersection-type.js

* Add test: intersection object type literal and type alias

Already passing!

* Add test: exported intersection between object type literals

Already passing

* Add test: intersection exported type aliases / type object literal

Already passing

* Add test: exported intersection between exported alias and literal

* Refactor

* Add support for intersection including imported type

This requires adding some run-time code which merges the
imported type into the existing types.

* Add some comments to the code

* Fix incorrect use of prop-types.any for intersection

The code generated to import a type from another file will
default to prop-types.any if the call to require() fails.
This will break if the imported type is used in an intersection,
because the generated code for that case will merge the
prop-types at runtime via Object.assign. For this case, we
check for prop-types.any and replace with '{}'.

* Add support for nested use of imported types

* Update comment re possible code duplication

* Fix intersection between two imported types

* Fix linting errors

* Remove stale comment

* Fix comment indent

* Add some tests for inline (non-alias) annotations

In an attempt to pinpoint some elusive weird behaviors, I created these
tests. They do pass.

* Passing test case for inline annotation with imported type

* Another passing test

* Another passing test

* Fix nested shape-intersect-runtime

* Add missing snapshot file

* Fix broken test

I fixed a bug where a boolean expression used || instead of == to determine the type,
which resulted in that condition always being true. Since that's fixed, we need to look
at a couple more types.

* Remove $debug invocations

* Remove obsolete comment

* Lint

* Add complex example for nested intersections with imported types

* Export objects, not shapes - shape-ify on import side

To properly handle intersections for imported types at runtime, we
need to export these types as objects, not shapes.

This is an intermediate commit, fixing the exports and only some
pieces of the import side while updating the tests.

Interesting, we still need to default to proptypes.any on failed
imports instead of {}. We don't know how to imported type would be
used, and proptypes.any will work in all contexts. For this reason,
we still need to check (as we do currently) for proptypes.any in
the runtime merge code, i.e. in the arguments to Object.assign.

Missing:

* Do not shape-ify variables in the runtime merge code (transpile time)
* Do not shape-ify functions (check at runtime)

* Do not generate shapes in runtime merge code

All parameters to Object.assign should be objects, not shapes. Fix
that for imported types.

* Convert objects to shape at runtime

We no longer import shapes, we import objects OR functions. For objects,
we have to convert to shapes at runtime. Note that for failed imports,
we always get propTypes.any; so this will continue to work as is.

* Add render tests

These tests render exemplary components and check if proptypes properly log
errors.

* Fix unnecessary shape() invocations

For chained intersections with imported types, we need to be careful
to use objects, not shapes, as argument to Object.assign. On the other
hand, we can't assign a shape to Foo.propTypes.

* Add another (non-render) test for deeply nested import intersections

* Add test for imported non-shape types

* Allow exporting for any type

* Fix check for object VS pre-defined proptype

We can't check if an imported type is an object - the check we had
always returns true. If we check of typeof(foo) === function, we
get the desired behavior: wrapping in a shape for objects and
pass-through for prop-types.

* Add documentation + refactor

* Lint.

* Add missing pieces from refactor