Fix functional component props (#88)
* Fix inline annotations for stateless functional components

This commit removes some superfluous checks: for GenericTypeAnnotations
with variables, we can just dispatch to convertNodeToPropTypes which
handles that case just fine. That function then returns the proper
intermediate representation with type 'raw'. This can be
passed as-is to makePropTypesAst by removin the superfluous check
for a 'properties' member.

* Fix non-object assignments to Foo.propTypes

We only ever assign objects to Foo.propTypes, not functions.
However, functions may be used when exporting a type declaration.

We now have two different methods to differentiate between these cases.