Allow `arrayOf`/`objectOf` elements to be required (#184)
* Add tests for `arrayOf`/`objectOf` required types

These tests currently fail.

wchargin-branch: arrayof-objectof-elements-required-tests

* Allow `arrayOf`/`objectOf` elements to be required

Summary:
Fixes #183; see rationale there.

Test Plan:
The code change to `convertToPropTypes` is sufficient to fix the tests
added and modified in the previous commit:

    yarn test objectof generic-array

Obviously, this is a wide-reaching change, so several other snapshots
were invalidated. I generated them, and manually inspected the results
to ensure that the changes are correct. They are, as far as I can tell;
the arrays of external types are harder to read, but by comparison with
surrounding code they seem okay to me.

wchargin-branch: arrayof-objectof-elements-required